% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{arma_optim}
\alias{arma_optim}
\title{Optimize arbitrary and differentiable functions}
\usage{
arma_optim(data = list(), evaluate, gradient,
  optimizer = optimzer_SA())
}
\arguments{
\item{data}{a named list of prior data you would like to supply to the evaluate function.}

\item{evaluate}{a function that is to be minimized. It should return a single numeric.}

\item{gradient}{optional, a function computing the gradient of \code{evaluate}}

\item{optimizer}{one of the many optimizers}
}
\description{
The function compiles the code to C++ and uses Armadillo and ensmallen to optimize it.
}
