% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compiler.R
\name{armacmp_compile}
\alias{armacmp_compile}
\title{Compile a function to C++}
\usage{
armacmp_compile(fun, function_name)
}
\arguments{
\item{fun}{a function}

\item{function_name}{the function name}
}
\value{
a list of type "armacmp_cpp_fun"
}
\description{
Compile a function to C++
}
