% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacmp.R
\name{compile}
\alias{compile}
\title{Compile Linear Algebra Code to C++}
\usage{
compile(fun, verbose = FALSE)
}
\arguments{
\item{fun}{a function}

\item{verbose}{print out compiler information}
}
\description{
Compile Linear Algebra Code to C++
}
