% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_DE}
\alias{optimizer_DE}
\title{Differential Evolution (DE)}
\usage{
optimizer_DE(
  populationSize = 100,
  maxGenerations = 2000,
  crossoverRate = 0.6,
  differentialWeight = 0.8,
  tolerance = 1e-05
)
}
\arguments{
\item{populationSize}{Scaling exponent for the step size.}

\item{maxGenerations}{Scaling exponent for evaluation step size.}

\item{crossoverRate}{Scaling parameter for step size.}

\item{differentialWeight}{Scaling parameter for evaluation step size.}

\item{tolerance}{Maximum absolute tolerance to terminate algorithm.}
}
\description{
Differential Evolution (DE)
}
