% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_GradientDescent}
\alias{optimizer_GradientDescent}
\title{Gradient Descent Optimizer}
\usage{
optimizer_GradientDescent(stepSize = 0.01, maxIterations = 1e+05,
  tolerance = 1e-05)
}
\arguments{
\item{stepSize}{Step size for each iteration}

\item{maxIterations}{Maximum number of iterations allowed (0 means no limit).}

\item{tolerance}{Maximum absolute tolerance to terminate algorithm.}
}
\description{
Gradient Descent Optimizer
}
