% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizers.R
\name{optimizer_SPSA}
\alias{optimizer_SPSA}
\title{Simultaneous Perturbation Stochastic Approximation (SPSA)}
\usage{
optimizer_SPSA(
  alpha = 0.602,
  gamma = 0.101,
  stepSize = 0.16,
  evaluationStepSize = 0.3,
  maxIterations = 1e+05,
  tolerance = 1e-05
)
}
\arguments{
\item{alpha}{Scaling exponent for the step size.}

\item{gamma}{Scaling exponent for evaluation step size.}

\item{stepSize}{Scaling parameter for step size.}

\item{evaluationStepSize}{Scaling parameter for evaluation step size.}

\item{maxIterations}{Maximum number of iterations allowed (0 means no limit).}

\item{tolerance}{Maximum absolute tolerance to terminate algorithm.}
}
\description{
Simultaneous Perturbation Stochastic Approximation (SPSA)
}
