% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armacmp.R
\name{armacmp}
\alias{armacmp}
\title{Compile Linear Algebra Code to C++}
\usage{
armacmp(code)
}
\arguments{
\item{code}{an R like expression that is translated to C++}
}
\description{
Compile Linear Algebra Code to C++
}
